// app/admin/_lib/tareaFuncion.type.ts
export interface FuncionConDetalles extends Omit<TareaFuncionBasePrisma, 'parametrosRequeridos' | 'orden'> {
    orden?: number; // orden es requerido en el estado local de la UI para DnD
    parametrosRequeridos?: (Omit<TareaFuncionParametroRequeridoBasePrisma, 'parametroRequerido'> & {
        parametroRequerido?: Pick<ParametroRequeridoBasePrisma, 'id' | 'nombreVisible' | 'nombreInterno' | 'tipoDato'> | null | undefined;
    })[];
    _count?: {
        tareas?: number;
    };
}